fix: cleanup

+++ app-scripts/install-edge.sh
@@ -1,7 +1,28 @@
 #!/bin/bash
-apt update && apt install -y curl
-curl https://packages.microsoft.com/keys/microsoft.asc | gpg --dearmor > microsoft.gpg
-install -o root -g root -m 644 microsoft.gpg /etc/apt/trusted.gpg.d/
-sh -c 'echo "deb [arch=amd64] https://packages.microsoft.com/repos/edge stable main" > /etc/apt/sources.list.d/microsoft-edge-dev.list'
-rm microsoft.gpg
-apt update && apt install -y microsoft-edge-stable
+app=microsoft-edge
+script=install-$app
+output=intune-script-logs/$script-script-output.txt
+log=/home/$(id -u -n 1000)/$output
+
+if ! (id -u -n 1000);
+then
+        mkdir -p /tmp/intune-script-logs
+        echo "user with id 1000 does not exist!" >> /tmp/$output
+        exit 1
+fi
+mkdir -p /home/$(id -u -n 1000)/intune-script-logs
+echo $(date -Iseconds) >> $log
+echo "hello-from-intune: $script" >> $log
+apt-get update
+if [ "$(dpkg -l | awk '/microsoft-edge/ {print }'|wc -l)" -ge 1 ]; then
+  apt update && apt install -y curl
+  curl https://packages.microsoft.com/keys/microsoft.asc | gpg --dearmor > microsoft.gpg
+  install -o root -g root -m 644 microsoft.gpg /etc/apt/trusted.gpg.d/
+  sh -c 'echo "deb [arch=amd64] https://packages.microsoft.com/repos/edge stable main" > /etc/apt/sources.list.d/microsoft-edge-dev.list'
+  rm microsoft.gpg
+  apt update && apt install -y microsoft-edge-stable
+  echo "$app already installed" >> $log
+else
+  echo "$app successfully installed" >> $log
+fi
+echo -e "bye-from-intune\n" >> $log
